Rename CurrentConnectionsCountAdvice and TotalRequestsCountAdvice (#23451)
* Move CurrentConnectionCountAdvice and RequestCountAdvice to proxy package

* Rename CurrentConnectionsCountAdvice and TotalRequestsCountAdvice
